requirements.txt
numpy>=1.26
scipy>=1.13
librosa>=0.10.2
soundfile>=0.12.1
faiss-cpu>=1.8.0
